+++ js/main.js
@@ -1,5 +1,5 @@
-import './loading-form.js';
-import './validation.js';
 import { cardsArray } from './cards-data.js';
 import { renderGallery } from './gallery.js';
+import './validation-loading-form.js';
+
 renderGallery(cardsArray());
